requirements.txt
matplotlib==3.8.4
mlflow==2.12.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
wandb==0.17.0
weave==0.50.2

